Move & rename subscriber, update dependencies
